# Parallel dual-prime 1024-bit modular exponentiation for AVX512_IFMA capable processors.