@kroeckx 2020-12-30
So you mention this is for IceLake processors, which seem to be a mobile version. It was my understanding that the use of those instructions will cause license-based downclocking. At least for some (older) generation Intel processors, this down clocking is probably not what we want. But looking at IceLake information, the downlclocking seems reasonable.

So I have to wonder if this is going to affect older CPUs in a negative way, and why you do this for a mobile CPU that's probably not going to do that many RSA operations.