@amatyuko-intc 2020-12-30
> So I have to wonder if this is going to affect older CPUs in a negative way, and why you do this for a mobile CPU that's probably not going to do that many RSA operations.

The older CPUs are not affected by this optimization as it utilizes AVX512_IFMA ISA, which is initially appeared in IceLake microarchitecture. As to downclocking issues, we intentionally sticked to 256-bit width registers (AVX12VL) with less power license instead of full 512-bit ones which reduces such effect.

IceLake has not only mobile version, there is a ICX server as well which can benefit from this.