@t8m 2021-01-05
I am a little bit worried about adding another special code path for 2k RSA keys for just specific family of processors.