@ilya-albrekht 2021-01-06
> Will all feature Intel server processors have support for this?

It will be supported by IceLake server processors and beyond, and will provide benefits for multiple generations (just like ADX code path that was initially added for Broadwell based platforms). Intel can release only client CPU performance numbers, as server parts are not available on the market yet. I'd expect no less speedup than on mobile IceLake processor.