@vanc 2021-01-11
> > Will all feature Intel server processors have support for this?
> 
> It will be supported by IceLake server processors and beyond, and will provide benefits for multiple generations (just like ADX code path that was initially added for Broadwell based platforms). Intel can release only client CPU performance numbers, as server parts are not available on the market yet. I'd expect no less speedup than on mobile IceLake processor.

Tried with the patch on an IceLake server CPU running at 2.2G base clock, I can confirm that RSA 2048 signing performance improved over 100%.

Before the patch:
```
                           sign        verify      sign/s  verify/s
rsa 2048 bits 0.000942s 0.000028s **1061.8**  36232.2
```

After the patch:
```
                           sign        verify      sign/s  verify/s
rsa 2048 bits 0.000466s 0.000029s **2145.4** 34772.2
```

That's about 103% performance gain.