@kroeckx 2021-01-11
> Is there any decision about this PR?

There isn't any decision on to include it or not.

I will probably not spend time reviewing this before the 3.0
release, but that shouldn't prevent someone else from doing this.
Others might feel the same, so this might need to wait for 3.1.