@mattcaswell 2021-01-21
> @mattcaswell, thanks for the comments. All of them are addressed in the latest commit.

Thanks. I'd really like to see a review from someone competent to look at the assembler side of things (even someone who is not a committer) before I add my approval. Otherwise I am happy with this.