@amatyuko-intc 2021-03-17
> Hmm. There appears to be some relevant failures in the CIs.

Yes, seems like CI switched to another GCC since the last PR update (I suspect 9.x? not sure how to get exact version from CI info) and it started producing a warning. Looking into it.