@amatyuko-intc 2021-03-17
Made a tiny change to to_words52() function to make it a bit cleaner for compiler optimizer. Should be good now.