@mattcaswell 2021-03-18
Still good. No need to reset 24 hour timer for this I think.