@guidovranken 2021-04-15
@amatyuko-intc or OpenSSL team: I would like to test this function with my fuzzer (https://github.com/guidovranken/cryptofuzz). IF there are bugs in the code, I can probably quickly find them. Unfortunately I don't have the requisite CPU. Is anyone willing to either give me (temporary) SSH access to a system with this CPU, or run Cryptofuzz on their system (with my guidance)?