@guidovranken 2021-04-16
> @guidovranken, I would be glad to help you run the tool on IceLake CPU, it would be interesting to check this. Do you have some pre-work ready to have this function call in the fuzzer? And I appreciate any guidance (you can find my email in the github profile).

Great! I will e-mail you.